refactor:post feed in userprofile
-moved posts, comments fetch for another component
